+++ packages/commonwealth/client/scripts/views/SublayoutHeader.tsx
@@ -7,7 +7,7 @@ import { CWCommunityAvatar } from './components/component_kit/cw_community_avata
 import { CWDivider } from './components/component_kit/cw_divider';
 import { CWIconButton } from './components/component_kit/cw_icon_button';
 import { isWindowSmallInclusive } from './components/component_kit/helpers';
-import { LoginSelector } from './components/header/login_selector';
+import { LoginSelector } from './components/Header/LoginSelector';
 import { CreateContentPopover } from './menus/create_content_menu';
 import { NotificationsMenuPopover } from './menus/notifications_menu';
 import { SearchBar } from './pages/search/search_bar';
