+++ packages/commonwealth/client/scripts/views/components/Header/LoginSelector.tsx
@@ -1,43 +1,34 @@
 import React, { useState, useEffect } from 'react';
 import ClickAwayListener from '@mui/base/ClickAwayListener';
 
-import { initAppState } from 'state';
-import { ChainBase, ChainNetwork, WalletId } from 'common-common/src/types';
+import { ChainBase, ChainNetwork } from 'common-common/src/types';
 import { addressSwapper } from 'utils';
-import $ from 'jquery';
+import { redraw } from 'mithrilInterop';
 
 import _ from 'lodash';
 
-import 'components/header/login_selector.scss';
+import 'components/Header/LoginSelector.scss';
 import {
   linkExistingAddressToChainOrCommunity,
   setActiveAccount,
 } from 'controllers/app/login';
-import { notifySuccess } from 'controllers/app/notifications';
-import { isSameAccount, pluralize } from 'helpers';
-import { setDarkMode } from 'helpers/darkMode';
+import { isSameAccount } from 'helpers';
 
 import app from 'state';
 import { User } from 'views/components/user/user';
 import { LoginModal } from 'views/modals/login_modal';
-import { FeedbackModal } from 'views/modals/feedback_modal';
-import WebWalletController from '../../../controllers/app/web_wallets';
-import Account from '../../../models/Account';
 import AddressInfo from '../../../models/AddressInfo';
 import ITokenAdapter from '../../../models/ITokenAdapter';
-import { SelectAddressModal } from '../../modals/select_address_modal';
 import { CWButton } from '../component_kit/cw_button';
 import { CWIconButton } from '../component_kit/cw_icon_button';
-import { CWText } from '../component_kit/cw_text';
-import { CWToggle, toggleDarkMode } from '../component_kit/cw_toggle';
 import { AccountSelector } from '../component_kit/cw_wallets_list';
 import { isWindowMediumSmallInclusive } from '../component_kit/helpers';
-import { UserBlock } from '../user/user_block';
-import { CWDivider } from '../component_kit/cw_divider';
 import { Popover, usePopover } from '../component_kit/cw_popover/cw_popover';
 import { Modal } from '../component_kit/cw_modal';
-import { useCommonNavigate } from 'navigation/helpers';
 import useForceRerender from 'hooks/useForceRerender';
+import { LoginSelectorMenuLeft } from 'views/components/Header/LoginSelectorMenuLeft';
+import { LoginSelectorMenuRight } from 'views/components/Header/LoginSelectorMenuRight';
+import { TOSModal } from 'views/components/Header/TOSModal';
 
 const CHAINBASE_SHORT = {
   [ChainBase.CosmosSDK]: 'Cosmos',
@@ -52,251 +43,7 @@ const CHAINNETWORK_SHORT = {
   [ChainNetwork.Terra]: 'Terra',
 };
 
-type LoginSelectorMenuLeftAttrs = {
-  activeAddressesWithRole: Array<Account>;
-  nAccountsWithoutRole: number;
-};
-
-export const LoginSelectorMenuLeft = ({
-  activeAddressesWithRole,
-  nAccountsWithoutRole,
-}: LoginSelectorMenuLeftAttrs) => {
-  const navigate = useCommonNavigate();
-  const forceRerender = useForceRerender();
-
-  const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
-  const [isSelectAddressModalOpen, setIsSelectAddressModalOpen] =
-    useState(false);
-  const [profileId, setProfileId] = useState(null);
-  const [selectedAddress, setSelectedAddress] = useState(null);
-
-  const changeSelectedAddress = () => {
-    const activeAccount = app.user.activeAccount ?? app.user.addresses[0];
-    setSelectedAddress(activeAccount.address);
-    forceRerender();
-  };
-
-  useEffect(() => {
-    // force rerender when new address is connected
-    app.user.isFetched.on('redraw', () => {
-      changeSelectedAddress();
-    });
-
-    return () => {
-      app.user.isFetched.off('redraw', () => {
-        changeSelectedAddress();
-      });
-    };
-  }, []);
-
-  useEffect(() => {
-    const activeAccount = app.user.activeAccount ?? app.user.addresses[0];
-    const chain =
-      typeof activeAccount.chain === 'string'
-        ? activeAccount.chain
-        : activeAccount.chain?.id;
-    const profile = app.newProfiles.getProfile(chain, activeAccount.address);
-    setProfileId(profile.id);
-    setSelectedAddress(activeAccount.address);
-  }, []);
-
-  const { activeAccounts } = app.user;
-
-  return (
-    <div className="LoginSelectorMenu left">
-      {app.activeChainId() && (
-        <>
-          <CWText type="caption" className="title">
-            Select address to use
-          </CWText>
-          {activeAccounts.map((account, i) => {
-            return (
-              <div
-                key={i}
-                className={`login-menu-item ${
-                  selectedAddress === account.address ? 'selected' : ''
-                }`}
-                onClick={async () => {
-                  await setActiveAccount(account);
-                  setSelectedAddress(account.address);
-                }}
-              >
-                <UserBlock
-                  user={account}
-                  selected={isSameAccount(account, app.user.activeAccount)}
-                  showRole={false}
-                  compact
-                  hideAvatar
-                />
-              </div>
-            );
-          })}
-        </>
-      )}
-      {activeAccounts.length > 0 && <CWDivider />}
-      <div
-        className="login-menu-item"
-        onClick={() => {
-          navigate(`/profile/id/${profileId}`, {}, null);
-        }}
-      >
-        <CWText type="caption">View profile</CWText>
-      </div>
-      <div
-        className="login-menu-item"
-        onClick={() => {
-          navigate(`/profile/edit`, {}, null);
-        }}
-      >
-        <CWText type="caption">Edit profile</CWText>
-      </div>
-      <div
-        className="login-menu-item"
-        onClick={() => {
-          if (nAccountsWithoutRole > 0) {
-            setIsSelectAddressModalOpen(true);
-          } else {
-            setIsLoginModalOpen(true);
-          }
-        }}
-      >
-        <CWText type="caption">
-          {nAccountsWithoutRole > 0
-            ? `${pluralize(nAccountsWithoutRole, 'other address')}...`
-            : 'Connect a new address'}
-        </CWText>
-      </div>
-      <Modal
-        content={
-          <SelectAddressModal
-            onModalClose={() => setIsSelectAddressModalOpen(false)}
-          />
-        }
-        onClose={() => setIsSelectAddressModalOpen(false)}
-        open={isSelectAddressModalOpen}
-      />
-      <Modal
-        content={<LoginModal onModalClose={() => setIsLoginModalOpen(false)} />}
-        isFullScreen={isWindowMediumSmallInclusive(window.innerWidth)}
-        onClose={() => setIsLoginModalOpen(false)}
-        open={isLoginModalOpen}
-      />
-    </div>
-  );
-};
-
-interface LoginSelectorMenuRightProps {
-  onLogout: () => void;
-}
-
-export const LoginSelectorMenuRight = ({
-  onLogout,
-}: LoginSelectorMenuRightProps) => {
-  const navigate = useCommonNavigate();
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isDarkModeOn, setIsDarkModeOn] = useState<boolean>(
-    localStorage.getItem('dark-mode-state') === 'on'
-  );
-
-  const resetWalletConnectSession = async () => {
-    /**
-     * Imp to reset wc session on logout as subsequent login attempts fail
-     */
-    const walletConnectWallet = WebWalletController.Instance.getByName(
-      WalletId.WalletConnect
-    );
-    await walletConnectWallet.reset();
-  };
-
-  return (
-    <>
-      <div className="LoginSelectorMenu right">
-        <div
-          className="login-menu-item"
-          onClick={() => navigate('/notification-settings', {}, null)}
-        >
-          <CWText type="caption">Notification settings</CWText>
-        </div>
-        <div className="login-menu-item">
-          <CWToggle
-            checked={isDarkModeOn}
-            onChange={(e) => {
-              isDarkModeOn
-                ? toggleDarkMode(false, setIsDarkModeOn)
-                : toggleDarkMode(true, setIsDarkModeOn);
-              e.stopPropagation();
-            }}
-          />
-          <div className="login-darkmode-label">
-            <CWText type="caption">Dark mode</CWText>
-          </div>
-        </div>
-        <CWDivider />
-        <div className="login-menu-item" onClick={() => setIsModalOpen(true)}>
-          <CWText type="caption">Send feedback</CWText>
-        </div>
-        <div
-          className="login-menu-item"
-          onClick={() => {
-            $.get(`${app.serverUrl()}/logout`)
-              .then(async () => {
-                await initAppState();
-                await resetWalletConnectSession();
-
-                notifySuccess('Logged out');
-                onLogout();
-                setDarkMode(false);
-              })
-              .catch(() => {
-                // eslint-disable-next-line no-restricted-globals
-                location.reload();
-              });
-          }}
-        >
-          <CWText type="caption">Logout</CWText>
-        </div>
-      </div>
-      <Modal
-        content={<FeedbackModal onModalClose={() => setIsModalOpen(false)} />}
-        onClose={() => setIsModalOpen(false)}
-        open={isModalOpen}
-      />
-    </>
-  );
-};
-
-type TOSModalProps = {
-  onAccept: () => void;
-  onModalClose: () => void;
-};
-
-const TOSModal = ({ onModalClose, onAccept }: TOSModalProps) => {
-  return (
-    <div className="TOSModal">
-      <div className="close-button-wrapper">
-        <CWIconButton
-          iconButtonTheme="primary"
-          iconName="close"
-          iconSize="small"
-          className="close-icon"
-          onClick={onModalClose}
-        />
-      </div>
-      <div className="content-wrapper">
-        <CWText>
-          By clicking accept you agree to the community's Terms of Service
-        </CWText>
-        <CWButton onClick={onAccept} label="Accept" />
-      </div>
-    </div>
-  );
-};
-
-type LoginSelectorProps = {
-  onJoinSuccess: () => void;
-};
-
-export const LoginSelector = ({ onJoinSuccess }: LoginSelectorProps) => {
+export const LoginSelector = () => {
   const forceRerender = useForceRerender();
   const [profileLoadComplete, setProfileLoadComplete] = useState(false);
   const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
@@ -339,13 +86,6 @@ export const LoginSelector = ({ onJoinSuccess }: LoginSelectorProps) => {
       </>
     );
   }
-
-  const activeAddressesWithRole = app.user.activeAccounts.filter((account) => {
-    return app.roles.getRoleInCommunity({
-      account,
-      chain: app.activeChainId(),
-    });
-  });
 
   const activeAccountsByRole = app.roles.getActiveAccountsByRole();
 
@@ -484,7 +224,7 @@ export const LoginSelector = ({ onJoinSuccess }: LoginSelectorProps) => {
         if (app.chain && ITokenAdapter.instanceOf(app.chain)) {
           await app.chain.activeAddressHasToken(app.user.activeAccount.address);
         }
-        onJoinSuccess(); // this triggers a state update from the parent to update the sibling component
+        redraw();
       } catch (err) {
         console.error(err);
       }
@@ -557,7 +297,6 @@ export const LoginSelector = ({ onJoinSuccess }: LoginSelectorProps) => {
               <Popover
                 content={
                   <LoginSelectorMenuLeft
-                    activeAddressesWithRole={activeAddressesWithRole}
                     nAccountsWithoutRole={nAccountsWithoutRole}
                   />
                 }
