+++ packages/commonwealth/client/styles/components/Header/LoginSelector.scss
@@ -1,0 +1,50 @@
+@import '../../shared';
+
+.LoginSelector {
+  display: flex;
+
+  .join-button-container {
+    border-top: 1px solid $neutral-300;
+    border-left: 1px solid $neutral-300;
+    border-bottom: 1px solid $neutral-300;
+  }
+
+  .button-container {
+    display: flex;
+
+    .left-button {
+      align-items: center;
+      border-top: 1px solid $neutral-300;
+      border-left: 1px solid $neutral-300;
+      border-bottom: 1px solid $neutral-300;
+      cursor: pointer;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      padding: 4px 8px;
+    }
+
+    .User {
+      display: flex;
+      align-items: center;
+
+      .user-display-name {
+        font-weight: 700;
+      }
+    }
+  }
+
+  .button-container {
+    display: flex;
+
+    .right-button {
+      align-items: center;
+      border: 1px solid $neutral-300;
+      cursor: pointer;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      padding: 4px 8px;
+    }
+  }
+}
