+++ packages/commonwealth/client/styles/components/Header/TOSModal.scss
@@ -1,0 +1,24 @@
+.TOSModal {
+  display: flex;
+  flex-direction: column;
+  height: 140px;
+  margin: 8px;
+  width: 400px;
+
+  .close-button-wrapper {
+    display: flex;
+    justify-content: flex-end;
+  }
+
+  .content-wrapper {
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+    margin-top: 4px;
+
+    .Text {
+      text-align: center;
+    }
+  }
+}
