--- packages/commonwealth/client/styles/components/header/login_selector.scss
@@ -1,128 +1,0 @@
-@import '../../shared';
-
-.LoginSelector {
-  display: flex;
-
-  .join-button-container {
-    border-top: 1px solid $neutral-300;
-    border-left: 1px solid $neutral-300;
-    border-bottom: 1px solid $neutral-300;
-  }
-
-  .button-container {
-    display: flex;
-
-    .left-button {
-      align-items: center;
-      border-top: 1px solid $neutral-300;
-      border-left: 1px solid $neutral-300;
-      border-bottom: 1px solid $neutral-300;
-      cursor: pointer;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      padding: 4px 8px;
-    }
-
-    .User {
-      display: flex;
-      align-items: center;
-
-      .user-display-name {
-        font-weight: 700;
-      }
-    }
-  }
-
-  .button-container {
-    display: flex;
-
-    .right-button {
-      align-items: center;
-      border: 1px solid $neutral-300;
-      cursor: pointer;
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      padding: 4px 8px;
-    }
-  }
-}
-
-.TOSModal {
-  display: flex;
-  flex-direction: column;
-  height: 140px;
-  margin: 8px;
-  width: 400px;
-
-  .close-button-wrapper {
-    display: flex;
-    justify-content: flex-end;
-  }
-
-  .content-wrapper {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    margin-top: 4px;
-
-    .Text {
-      text-align: center;
-    }
-  }
-}
-
-.LoginSelectorMenu {
-  background-color: $white;
-  border-radius: $border-radius-corners;
-  box-shadow: $elevation-3;
-  position: relative;
-
-  &.left {
-    right: 16px;
-
-    .UserBlock {
-      gap: 24px;
-    }
-  }
-
-  &.right {
-    right: -8px;
-  }
-
-  .title {
-    padding: 4px 16px;
-    color: $neutral-500;
-  }
-
-  .login-menu-item {
-    cursor: pointer;
-    display: flex;
-    gap: 8px;
-    padding: 8px 16px;
-    width: 100%;
-
-    &:not(.selected) {
-      svg {
-        fill: transparent;
-      }
-
-      &:hover {
-        svg {
-          fill: $neutral-500;
-        }
-      }
-    }
-
-    &:hover {
-      background-color: $neutral-100;
-    }
-  }
-
-  .login-darkmode-label {
-    display: grid;
-    align-items: center;
-  }
-}
