+++ packages/commonwealth/client/styles/menus/notifications_menu.scss
@@ -1,4 +1,4 @@
-@import '../../shared';
+@import '../shared';
 
 .unreads-icon svg {
   cursor: pointer;
